Replace calculate_root with calculate_root2
